% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{BatchTrue2ObservedCounts}
\alias{BatchTrue2ObservedCounts}
\title{Simulate observed count matrix given technical biases and the true counts in multiple batches}
\usage{
BatchTrue2ObservedCounts(true_counts, protocol, batch_effect = 0.1,
  nbatch = 3, alpha_mean = 0.1, alpha_sd = 0.02, lenslope = 0.01,
  nbins = 20, amp_bias_limit = c(-0.2, 0.2), rate_2PCR = 0.8, nPCR = 18,
  depth_mean, depth_sd)
}
\arguments{
\item{true_counts}{output of SimulateTrueCounts, with true count matrix and cell meta data}

\item{protocol}{a string, can be "ss2" or "umi"}

\item{batch_effect}{the ratio of standard deviation to mean when sampling per batch parameters, default is 0.1}

\item{nbatch}{number of batches, default is 3}

\item{alpha_mean}{the mean of rate of subsampling of transcripts during capture step, default at 10% efficiency}

\item{alpha_sd}{the std of rate of subsampling of transcripts}

\item{lenslope}{amount of length bias}

\item{nbins}{number of bins for gene length}

\item{amp_bias_limit}{range of amplification bias for each gene, a vector of length ngenes}

\item{rate_2PCR}{PCR efficiency, usually very high, default is 0.8}

\item{nPCR}{the number of PCR cycles, default is 16}

\item{depth_mean}{mean of sequencing depth}

\item{depth_sd}{std of sequencing depth}
}
\description{
Simulate observed count matrix given technical biases and the true counts in multiple batches
}
