% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{BestMatchParams}
\alias{BestMatchParams}
\title{Get the best matched parameter}
\usage{
BestMatchParams(tech, counts, plotfilename)
}
\arguments{
\item{tech}{'ss2','umi' the match database are constructed based on reasonable values for each technology}

\item{counts}{expression matrix}

\item{plotfilename}{output name for qqplot}
}
\value{
three set of best matching parameters that was used to simulate the best matching dataset to the experimental dataset
}
\description{
This function matches a real dataset to a database of summary information of simulated datasets, plots a qqplot for user to visualize similarity between their dataset and the simulated dataset, and suggests parameters to use in the simulation
}
