% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{cal_ARI}
\alias{cal_ARI}
\title{calculate adjusted rand index between labels from a clustering algorithm and known labels}
\usage{
cal_ARI(cluster_res, label)
}
\arguments{
\item{cluster_res}{results from a cluster algorithm}

\item{label}{known label of the cells}
}
\description{
calculate adjusted rand index between labels from a clustering algorithm and known labels
}
