% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{ClusterQuality}
\alias{ClusterQuality}
\title{cluster-ability of a dataset}
\usage{
ClusterQuality(data, meta, n_pc, perplexity, dims, npop = 5,
  return_kmeans = F, pca_scale = F)
}
\arguments{
\item{data:}{transcriptomics matrix}

\item{meta:}{information about cells, must contain 'pop' column}

\item{n_pc:}{number of principal components that is fed to tsne}

\item{perplexity:}{perplexity parameter for tsne}

\item{dims:}{number of dimensions kept after tsne}
}
\description{
how well can the cell population be reconstructed by dimensionality reduction methods
}
