% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{cor_pseudotime}
\alias{cor_pseudotime}
\title{calculate pseudotime correlation between our true pseudotime and predicted pseudotime
true pseudotime is a list where each element corresponds to an lineage}
\usage{
cor_pseudotime(true_time, pred_time)
}
\description{
calculate pseudotime correlation between our true pseudotime and predicted pseudotime
true pseudotime is a list where each element corresponds to an lineage
}
