% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{DE_EVF}
\alias{DE_EVF}
\title{Generating EVFs that is differential in one population and not in all other populations}
\usage{
DE_EVF(de_pop, de_evf_mean, cell_pop, Sigma, nevf, seed)
}
\arguments{
\item{de_pop}{the population that has a differential DE, it can be either a list (each element of the list is the populations that should be DE) of a vector}

\item{de_evf_mean}{the mean of evf of the DE population (mean for all other populations are zero), it can be either a list or a vector. If it is a list the de_pop also have to be the same length list and each element in the de_pop have the same length as each element in the evf_mean list}

\item{cell_pop}{the population id for each cell from the discrete evf simulation}

\item{Sigma}{The standard deviation of the brownian motion of EVFs changing along the tree}

\item{nevf}{Number of EVFs per cell (taht are DE)}
}
\value{
a list of two object, one is the evf, and the other is a dataframe indicating the population each cell comes from (pop)
}
\description{
Generating EVFs that is differential in one population and not in all other populations
}
