% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{Dec2k_based}
\alias{Dec2k_based}
\title{convert a decimal number to k based number}
\usage{
Dec2k_based(k, nbits, dec_number)
}
\arguments{
\item{k}{the base number}

\item{nbits}{the length (number of bits) in the target string}

\item{dec_number}{the decimal number to convert}
}
\value{
a vector with values as the converted string at each position
}
\description{
convert a decimal number to k based number
}
