% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_functions.R
\name{deseq}
\alias{deseq}
\title{RPM normalization}
\usage{
deseq(x)
}
\arguments{
\item{x}{matrix of transcript counts where rows are genes and columns are cells}
}
\value{
list(s,counts) s is a ncells long vector that stores the scaling factors for each cell, counts is the gene-cell matrix after normalization
}
\description{
This function runs the DESeq normalization (from Vallejos2017 supplementary material), DESeq defines scaling factor estimates based on a pseudoreference sample, which is built with the geometric mean of gene counts across all cells.
}
