% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{dist2diag}
\alias{dist2diag}
\title{how well do dots fit to the diagonal (for evaluating our qqplots)}
\usage{
dist2diag(x, y, nbins)
}
\arguments{
\item{x}{a vector on x axis}

\item{y}{a vector on y axis with the same length as x}
}
\description{
how well do dots fit to the diagonal (for evaluating our qqplots)
}
