% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{DivideBatches}
\alias{DivideBatches}
\title{Divide the observed counts into multiple batches by adding batch effect to each batch}
\usage{
DivideBatches(observed_counts_res, nbatch, batch_effect_size = 1)
}
\arguments{
\item{observed_counts_res}{the output from True2ObservedCounts}

\item{nbatch}{number of batches}

\item{batch_effect_size}{amount of batch effects. Larger values result in bigger differences between batches. Default is 1.}
}
\description{
Divide the observed counts into multiple batches by adding batch effect to each batch
}
