% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Dropped2Biased}
\alias{Dropped2Biased}
\title{Add Bias Terms and batch effects on the dropped count matrix}
\usage{
Dropped2Biased(dropped_exprs, nbins, randseed, gcslope, lenslope, batch,
  epsilon)
}
\arguments{
\item{dropped_exprs}{dropped expression matrix}

\item{nbins}{number of bins when simulating quantile bias}

\item{randseed}{random seed when assigning simulated genes to GC and length bins}

\item{gcslope}{the slope of the GC bias line (the middle category is 1)}

\item{lenslope}{the slope of the length bias line (middle category is 1)}

\item{batch}{a constant that increase or decreases the expression in the same batch by X fold}

\item{epsilon}{The standard deviation of the log error term that is multiplied to each gene}
}
\value{
biased_exprs The exrpession matrix
}
\description{
randomly sample GC and length content of the simulated genes, and bias the exrpession according to it. Also add batch effect that is applied to all the the cells in the same batch, and a random epsilon error term to every number
}
\examples{
Dropped2Biased()
}
