% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{evalImpulse}
\alias{evalImpulse}
\title{Simulate multiple discrete population in multiple batches}
\usage{
evalImpulse(vecImpulseParam, vecTimepoints)
}
\arguments{
\item{vecImpulseParam}{(numeric vector number of impulse model parameters)
\{beta, h0, h1, h2, t1, t2\}
Vector of impulse model parameters.}

\item{vecTimepoints}{(numeric vector length number of time points) 
Time points to be evaluated.}

\item{phyla}{a tree in the phylo format}

\item{nevf}{number of evfs}

\item{nbatches}{number of batches}

\item{evf_sd}{a vector of length nevf, the standard deviation of the normal distribution to sample evf from}

\item{ncells}{a matrix of size nbatch*npop, each row is a vector of length npop, and represents the number of cells from each population in each batch}

\item{randseed}{the random seed to generate the gc and length bins (should keep constant for each experiment regardless of the batch and population, but should change for replicates of simulations)}

\item{gene_effects}{a matrix of ngenes * nevf, generated by GeneEffects function}

\item{bimod}{the proportion of distance to move a point in the kon-koff space towards the center of the kon-koff space where the gene expression is bimodal}

\item{alpha}{a vector of length nbatches: library prep capture efficiency, or dropout of transcript: the number of of captured transcript is distributed as the number of successes with probability of success alpha in X trials, where X is the true number of transcripts}

\item{alpha_sd}{a vector of length nbatches: the standard deviation of alpha (to add nsimoise for different efficiency for each cell)}

\item{nbins}{the number of bins for gc and length bias}

\item{gcbias}{a vector of length nbatches: the magnitude of gc bias (a number between 0 and 2)}

\item{lenbias}{a vector of length nbatches: the magnitude of length bias (a number between 0 and 2)}

\item{batch}{a vector of length nbatches: the constant to multiply all transcripts by}

\item{noise}{a vector of length nbatches: the standard deviation of a normal distribution where the log(noise) is sampled from}

\item{sim_results}{output of sim1Pop1Batch, list a list of 4 elements: evf, true counts, sampled counts and biased counts.}
}
\value{
list a list of 2 elements: all_counts and meta. all_counts is a list of 3 matrices of nrow=number of genes and ncol=sum of number of cells in each population. The fist matrix is the real counts, the second matrix is the counts after dropout, and the third matrix is the biased counts. each row of meta correspond to 1 cell in the count matrix, and the most important information in it is the 'pop' column that tells which population a cell comes from and 'batch' the column that tells which batch a cell is from
Test results of scaling normalization

This function runs rpm, deseq, tmm, uq (quantile normalization), scran, and clt normalization

cor_compare a matrix of ngenes * 2*(1+nmethods), each row is a gene, and the columns are the correlation between the true counts to dropped counts, dropped counts corrected, biased counts and dropped counts corrected
Compute value of impulse function given parameters.

Compute value of impulse function given parameters.
Enforces lower bound on value of function to avoid numerical
errors during model fitting.

vecImpulseValue (vec number of vecTimepoints) 
Model values for given time points.
}
\description{
This function generates a evf variance-covariance matrix from a tree structure in the
}
\seealso{
Compiled version: \link{evalImpulse_comp}
}
\author{
David Sebastian Fischer
}
