% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{EVF2TrueCounts}
\alias{EVF2TrueCounts}
\title{Getting True Counts from EVF and Gene effects}
\usage{
EVF2TrueCounts(allparams, sim_master, evf, gene_effects, bimod)
}
\arguments{
\item{allparams}{the sets of parameter used to simulate master equations}

\item{sim_master}{the master equation results}

\item{evf}{a vector of length nevf, the cell specific extrinsic variation factor}

\item{gene_effects}{a list of three matrices (generated using the GeneEffects function), each corresponding to one kinetic parameter. Each matrix has nevf columns, and ngenes rows.}

\item{bimdod}{the parameter for increasing the bimodality of dataset. Takes value between 0 and 1, where 0 means keep the kinetic parameters as is, and 1 means move the kinetic parameters on the kon-koff plane to a space where the distribution of transcript counts is bimodal}
}
\value{
the simulated (true) expression matrix
}
\description{
This function first calls the Get_params function to calculate the kinetic parameters from evf and gene effects. 
And then tweaks the bimodality variable
At last finds the best matched simulation and sample from the master equation and outputs the true expression of a single cell
}
\examples{
EVF2TrueCounts()
}
