% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_len_pool}
\alias{gene_len_pool}
\title{a pool of gene lengths to sample from}
\format{a vector.}
\usage{
data(gene_len_pool)
}
\description{
a pool of gene lengths to sample from
}
\examples{
data(gene_len_pool)
}
\keyword{datasets}
