% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{GeneEffects}
\alias{GeneEffects}
\title{Getting GeneEffects matrices}
\usage{
GeneEffects(ngenes, nevf, randseed, prob, sd)
}
\arguments{
\item{ngenes}{number of genes}

\item{nevf}{number of evfs}

\item{prob}{the probability that the effect size is not 0}

\item{sd}{the standard deviation of the normal distribution where the non-zero effect sizes are dropped from}

\item{random}{seed (should produce same result if ngenes, nevf and randseed are all the same)}
}
\value{
a list of 3 matrices, each of dimension ngenes * nevf
}
\description{
This function randomly generates the effect size of each evf on the dynamic expression parameters
}
\examples{
GeneEffects()
}
