% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{getDEgenes}
\alias{getDEgenes}
\title{retrieve genes' differential expression information
it outputs three measures for every gene: 
nDiffEVF: the number of DiffEVFs used for each gene
logFC_theoretical: log2 fold change based on kinetic parameters
wil.p_true_counts: adjusted wilcoxon p-value based on true counts}
\usage{
getDEgenes(true_counts_res, popA, popB)
}
\arguments{
\item{true_counts_res}{the output of function SimulateTrueCounts()}

\item{popA}{the first population to be compared with (usually a number)}

\item{popB}{the second population to be compared with}
}
\description{
retrieve genes' differential expression information
it outputs three measures for every gene: 
nDiffEVF: the number of DiffEVFs used for each gene
logFC_theoretical: log2 fold change based on kinetic parameters
wil.p_true_counts: adjusted wilcoxon p-value based on true counts
}
