% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{getTrajectoryGenes}
\alias{getTrajectoryGenes}
\title{retrieve the information of cells on continuous trajectories
Outputs a data frame, where each row corresponds to a cell
Each cell has information "pseudotime" (distance from root) and "branch" (on which branch is the cell)}
\usage{
getTrajectoryGenes(cell_meta)
}
\arguments{
\item{cell_meta}{the cell meta information stored in the output of SimulateTrueCounts() or True2ObservedCounts()}
}
\description{
retrieve the information of cells on continuous trajectories
Outputs a data frame, where each row corresponds to a cell
Each cell has information "pseudotime" (distance from root) and "branch" (on which branch is the cell)
}
