% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{len2nfrag}
\alias{len2nfrag}
\title{from transcript length to number of fragments (for the nonUMI protocol)}
\format{a vector.}
\usage{
data(len2nfrag)
}
\description{
from transcript length to number of fragments (for the nonUMI protocol)
}
\examples{
data(len2nfrag)
}
\keyword{datasets}
