% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{len2prob3pri}
\alias{len2prob3pri}
\title{from transcript length to number of fragments (for the UMI protocol)}
\format{a vector.}
\usage{
data(len2prob3pri)
}
\description{
from transcript length to number of fragments (for the UMI protocol)
}
\examples{
data(len2prob3pri)
}
\keyword{datasets}
