% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{MasterEqn2}
\alias{MasterEqn2}
\title{Calculating Master Equation}
\usage{
MasterEqn2(rateParams)
}
\arguments{
\item{rateParams}{a vector of length 4 including kon, koff, s, d}
}
\value{
a vector of X where the probability of having less or equal to X transcripts is greater than 0.999
}
\description{
This function calculates the distribution of transcritp counts in single cells given the dynamic parameters for gene expression
}
\examples{
MasterEqn2(c(1,1,10,0.1))
}
