% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{match_params}
\alias{match_params}
\title{distribution of kinetic parameters learned from the Zeisel UMI cortex datasets}
\format{a data frame.}
\usage{
data(param_realdata.zeisel.imputed)
}
\description{
distribution of kinetic parameters learned from the Zeisel UMI cortex datasets
}
\examples{
data(param_realdata.zeisel.imputed)
}
\keyword{datasets}
