% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{MatchParams}
\alias{MatchParams}
\title{find closest matching simulated parameter sets}
\usage{
MatchParams(sim_master, allparams, log_count_bin, counts, plotting = TRUE,
  samplename)
}
\arguments{
\item{sim_master}{the output of the master equation simulation serving as the database}

\item{allparams}{the parameters that produced the sim_master}

\item{log_count_bin}{the log scale bins}

\item{counts}{the real expression matrix that we are trying to fit to}

\item{plotting}{(boolean) whether to generate plots or not, default is true}

\item{samplename}{the prefix of the output plots}
}
\value{
match_params the best matched parameters
}
\description{
using kd-tree method find the simulated distribution that is closest to the observed distribution, and as diagnostic, plot the expression heat map of the true counts and the best-match simulated counts, and also the distribution of the parameters
}
