% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{multiplot}
\alias{multiplot}
\title{arrange multiple plots from ggplot2 to one figure}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\description{
arrange multiple plots from ggplot2 to one figure
}
