% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_functions.R
\name{Normalize}
\alias{Normalize}
\title{Run all normalization}
\usage{
Normalize(x)
}
\arguments{
\item{x}{matrix of transcript counts where rows are genes and columns are cells}
}
\value{
list(sf,results) sf is a nmethods * ncells matrix that stores the scaling factors of different methods, results are the gene-cell matrix after normalization
}
\description{
This function runs rpm, deseq, tmm, uq (quantile normalization), scran, and clt normalization
}
