% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{Npop1Batch}
\alias{Npop1Batch}
\title{Simulate multiple discrete population in 1 batch}
\usage{
Npop1Batch(phyla, nevf, evf_sd, ncells, randseed, gene_effects, bimod, alpha,
  alpha_sd, nbins, gcbias, lenbias, batch, noise)
}
\arguments{
\item{phyla}{a tree in the phylo format}

\item{nevf}{number of evfs}

\item{evf_sd}{a vector of length nevf, the standard deviation of the normal distribution to sample evf from}

\item{ncells}{a vector of length npop, the number of cells in each population}

\item{randseed}{the random seed to generate the gc and length bins (should keep constant for each experiment regardless of the batch and population, but should change for replicates of simulations)}

\item{gene_effects}{a matrix of ngenes * nevf, generated by GeneEffects function}

\item{bimod}{the proportion of distance to move a point in the kon-koff space towards the center of the kon-koff space where the gene expression is bimodal}

\item{alpha}{library prep capture efficiency, or dropout of transcript: the number of of captured transcript is distributed as the number of successes with probability of success alpha in X trials, where X is the true number of transcripts}

\item{alpha_sd}{the standard deviation of alpha (to add nsimoise for different efficiency for each cell)}

\item{nbins}{the number of bins for gc and length bias}

\item{gcbias}{the magnitude of gc bias (a number between 0 and 2)}

\item{lenbias}{the magnitude of length bias (a number between 0 and 2)}

\item{batch}{the constant to multiply all transcripts by}

\item{noise}{the standard deviation of a normal distribution where the log(noise) is sampled from}
}
\value{
list a list of 2 elements: all_counts and meta. all_counts is a list of 3 matrices of nrow=number of genes and ncol=sum of number of cells in each population. The fist matrix is the real counts, the second matrix is the counts after dropout, and the third matrix is the biased counts. each row of meta correspond to 1 cell in the count matrix, and the most important information in it is the 'pop' column that tells which population a cell comes from
}
\description{
This function generates a evf variance-covariance matrix from a tree structure in the
}
