% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{percent_nonzero}
\alias{percent_nonzero}
\title{Calculate the percentage of non-zero values in a vector x}
\usage{
percent_nonzero(x)
}
\description{
Calculate the percentage of non-zero values in a vector x
}
