% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{Phyla3}
\alias{Phyla3}
\title{Creating an example tree with 3 tips}
\usage{
Phyla3(plotting = F)
}
\arguments{
\item{plotting}{True for plotting the tree on console, False for no plot}
}
\value{
a tree object
}
\description{
Creating an example tree with 3 tips
}
