% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{plot_nreads_UMI}
\alias{plot_nreads_UMI}
\title{plot the histogram of number of reads per UMI}
\usage{
plot_nreads_UMI(hist_res, plot_title)
}
\description{
plot the histogram of number of reads per UMI
}
