% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{plotFNR}
\alias{plotFNR}
\title{Plotting FNR}
\usage{
plotFNR(expr_matrix, data_name, ncols)
}
\arguments{
\item{expr_matrix}{expression matrix}

\item{data_name}{what to name the plot}

\item{ncols}{number of colors in the color ramp}
}
\description{
Make 2 plots: one is the FNR curves, and the other one is the barplots of AUC
}
