% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{plotFNRreal}
\alias{plotFNRreal}
\title{Plotting real data FNR}
\usage{
plotFNRreal(expr_matrix, hk, data_name, col_vec)
}
\arguments{
\item{expr_matrix}{expression matrix}

\item{hk}{house keeping genes}

\item{data_name}{title string}

\item{col_vec}{color vector}
}
\description{
Plotting real data FNR
}
