% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{plotFNRsim}
\alias{plotFNRsim}
\title{Plotting simulated FNR}
\usage{
plotFNRsim(current_counts, true_counts, x_use = "true", titlestr)
}
\arguments{
\item{current_counts}{expression matrix}

\item{true_counts}{true expression matrix}

\item{titlestr}{title string}
}
\description{
Plotting simulated FNR
}
