% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{PlotGCLENbias}
\alias{PlotGCLENbias}
\title{Plotting GC bias and Length}
\usage{
PlotGCLENbias(filename, outputname)
}
\arguments{
\item{filename}{the name of the simualted robj data file}
}
\description{
Plotting GC bias and Length
}
