% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{PlotParamHist}
\alias{PlotParamHist}
\title{Plotting the histograms of kon,koff,s values}
\usage{
PlotParamHist(params, samplename)
}
\arguments{
\item{params}{a matrix of 3 columns, the first one is kon, the second is koff and the third is s}

\item{samplename}{the prefix of the plot, the suffix is '.params_dist.jpeg'}
}
\value{
make a plot of three histograms
}
\description{
plot colored histograms of parameters
}
