% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{plotPCAbasic}
\alias{plotPCAbasic}
\title{Plotting PCA results (PC1 and PC2)}
\usage{
plotPCAbasic(PCAres, col_vec, figuretitle)
}
\arguments{
\item{PCAres}{the PCA results}

\item{col_vec}{a vector to specify the colors for each point}

\item{figuretitle}{title for the plot}
}
\description{
Plotting PCA results (PC1 and PC2)
}
