% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{PlotTsne}
\alias{PlotTsne}
\title{Perform tSNE}
\usage{
<<<<<<< HEAD
PlotTsne(meta, data, plotname, label, n_pc = 20, evf_type = "discrete",
  saving = F, perplexity = 10)
=======
PlotTsne(meta, data, plotname, label, evf_type = "discrete", saving = F,
  perplexity = 10)
>>>>>>> b1c727f6ac83412053a415eccd221e3a5179bbd7
}
\arguments{
\item{meta:}{simulation parameters}

\item{data:}{expression matrix}

\item{plotname:}{the name of the jpeg file}

\item{label:}{the column name of the meta data that the points needs to be colored by}
}
\description{
This function takes the ground truth (meta), the expresssion matrix (data), and performs tSNE, saves the result to a jpeg file with user specified name
}
