% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{QQplot2RealData}
\alias{QQplot2RealData}
\title{function QQplot2RealData; plot both with ggplot2 and with basic plot}
\usage{
QQplot2RealData(real_data, sim_data, expressed_prop, plot_file_name,
  print_data_dir = NA)
}
\arguments{
\item{real_data}{gene expression matrix of experimental data}

\item{sim_data}{gene expression matrix of simulated data}

\item{plot_file_name}{the pdf file name for the output plots}

\item{print_data_dir}{the directory to print the source data for plots; if NA then do not print}

\item{express_prop}{the minimum proportion of cells a gene should be expressed in}
}
\description{
function QQplot2RealData; plot both with ggplot2 and with basic plot
}
