% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{rand.index}
\alias{rand.index}
\title{rand index}
\usage{
rand.index(group1, group2)
}
\arguments{
\item{group1}{first clustering}

\item{group2}{second clustering}
}
\description{
compare two clustering result: proportion of pairs of individual that share the same clustering result in both groupings
}
