% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{rescale2range}
\alias{rescale2range}
\title{rescale2range}
\usage{
rescale2range(vec, n)
}
\arguments{
\item{vec}{input vector}

\item{n}{the largest integer to scale the vector to (the smallest is 1)}
}
\description{
Subfunction for Plotting FNR. Rescale the values in vec such that the lagest is n, and the smallest is 1.
}
