% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_functions.R
\name{rpm}
\alias{rpm}
\title{RPM normalization}
\usage{
rpm(x)
}
\arguments{
\item{x}{matrix of transcript counts where rows are genes and columns are cells}
}
\value{
list(s,counts) s is a ncells long vector that stores the scaling factors for each cell, counts is the gene-cell matrix after normalization
}
\description{
This function runs the Reads Per Million normalization (from Vallejos2017 supplementary material), so the total read counts per cell is made equal across all cells
}
