% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SampleAlpha}
\alias{SampleAlpha}
\title{Generate alpha vector for ncells}
\usage{
SampleAlpha(ncells, alpha, alpha_sd)
}
\arguments{
\item{ncells}{number of cells}

\item{alpha}{the mean value of a normal distribution where alpha is sampled from}

\item{alpha_sd}{the standard deviation value of a normal distribution where alpha is sampled from}
}
\value{
a vector with an alpha value for each cell
}
\description{
Generate alpha vector for ncells
}
