% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SampleBatch}
\alias{SampleBatch}
\title{Generate gene bias vector for each gene}
\usage{
SampleBatch(ngenes, nbatch, batch_sd)
}
\arguments{
\item{ngenes}{number of gene}

\item{nbatch}{number of bathces}

\item{batch_sd}{the standard deviation for the random per genebias in each batch (same for each batch)}
}
\value{
a vector with an batch bias for each gene
}
\description{
Generate gene bias vector for each gene
}
