% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SampleBias}
\alias{SampleBias}
\title{Generate gene bias vector for each gene}
\usage{
SampleBias(ngenes, slope, nbins, randseed)
}
\arguments{
\item{ngenes}{number of gene}

\item{slope}{a float value between -2 and 2 determining the strength of bias}

\item{nbins}{number of bins for gene biases}

\item{randseed}{random seed used for assigning genes into different bins}
}
\value{
a vector with a GC or Length bias for each gene
}
\description{
Generate gene bias vector for each gene
}
