% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SampleExprs}
\alias{SampleExprs}
\title{Sample read counts from master equation}
\usage{
SampleExprs(sim_dist, ncells)
}
\arguments{
\item{sim_dist}{master equation}

\item{cells}{the number of times to sample from each distribution}
}
\value{
counts is a matrix of dimension ngenes * ncells
}
\description{
This function takes ngenes distributions of transcript counts calculated by the master equation and sample from them
}
\examples{
SampleExprs()
}
