% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim1Pop1Batch}
\alias{sim1Pop1Batch}
\title{Simulate 1 homogenous population in 1 batch}
\usage{
sim1Pop1Batch(evf_mean, evf_sd, ncells, randseed, gene_effects, bimod, alpha,
  alpha_sd, nbins, gcbias, lenbias, batch, noise, matched_params)
}
\arguments{
\item{evf_mean}{a vector of length nevf, the means of the normal distribution to sample evf from}

\item{evf_sd}{a vector of length nevf, the standard deviation of the normal distribution to sample evf from}

\item{ncells}{a integer of the number of cells to simulate}

\item{randseed}{the random seed to generate the gc and length bins (should keep constant for each experiment regardless of the batch and population, but should change for replicates of simulations)}

\item{gene_effects}{a matrix of ngenes * nevf, generated by GeneEffects function}

\item{bimod}{the proportion of distance to move a point in the kon-koff space towards the center of the kon-koff space where the gene expression is bimodal}

\item{alpha}{library prep capture efficiency, or dropout of transcript: the number of of captured transcript is distributed as the number of successes with probability of success alpha in X trials, where X is the true number of transcripts}

\item{alpha_sd}{the standard deviation of alpha (to add nsimoise for different efficiency for each cell)}

\item{nbins}{the number of bins for gc and length bias}

\item{gcbias}{the magnitude of gc bias (a number between 0 and 2)}

\item{lenbias}{the magnitude of length bias (a number between 0 and 2)}

\item{noise}{the standard deviation of a normal distribution where the log(noise) is sampled from}

\item{batch_mean}{The mean batch effect from which the batch effect of each gene is sampled from}

\item{batch_sd}{The standard deviation of batch effect from which the batch effect of each gene is sampled from}
}
\value{
list a list of 4 elements: evf, true counts, sampled counts and biased counts.
}
\description{
This function takes a gene effect matrix, the parameters for sampling extrinsic, intronsic and experimental variation, and outputs results of simulation
}
