% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{Sim_LogDist}
\alias{Sim_LogDist}
\title{Get the logged distribution from master equation simulations}
\usage{
Sim_LogDist(dist, log_count_bins)
}
\arguments{
\item{dist}{a list of master equation simulation results, each element is a vector of length K}

\item{log_count_bins}{a vector of form seq(min,max,stepsize), or doesn't have equal distance bins}
}
\value{
a matrix where each column is a bin, and each row is one distribution, and the contents are frequencies of probability of being in each bin
}
\description{
This function converts the frequency on integers from (0-K transcripts) to log scaled frequency, where the log_count_bins gives the range for each count bin
}
