% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SimulateCounts}
\alias{SimulateCounts}
\title{Simulate count matrix from evf, gene effects and technical biases}
\usage{
SimulateCounts(ncells, ngenes, nevf, nbatches, evfs, gene_effects, alphas, beta,
  gcbias, lenbias, batch, batch_id, meta1, epsilon)
}
\arguments{
\item{ncells}{number of cells}

\item{ngenes}{number of genes}

\item{nevf}{number of evfs}

\item{nbatches}{number of batches}

\item{evfs}{matrix with ncells rows and nevf columns, can be generated by ContinuousEVF and DiscreteEVF}

\item{gene_effects}{a list of length 3, each element is a sparse matrix determining how each gene is affected by the EVF of that cell}

\item{alphas}{a vector of length ncells}

\item{beta}{a single float value between 0 and 1 determining the bimodality of entire dataset}

\item{gcbias}{a vector of length ngenes}

\item{lenbias}{a vector of length ngenes}

\item{batch}{a matrix of ngenes rows and nbatch columns}

\item{batch_id}{a vector of length ncells indicating which batch each cell comes from (randomly generated)}

\item{meta1}{meta data generated from EVF function (population, depth in the tree, etc.)}

\item{epsilon}{a single float value the magnitude of independent noise for each gene in each cell}
}
\value{
a list of 3 objects, one is a list of the count matrices (true counts, dropped counts and biased counts), the per cell meta data, and the per gene meta data
}
\description{
Simulate count matrix from evf, gene effects and technical biases
}
