% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SimulateDataset}
\alias{SimulateDataset}
\title{Simulate count matrix from evf, gene effects and technical biases}
\usage{
SimulateDataset(ncells, ngenes, nbatch, nevf, Sigma = 0.5, phyla,
  alpha = 0.1, alpha_sd = 0.01, nbins = 10, gcslope = 0.1,
  lenslope = -1, batch_sd = 0.5, beta = 0.5, epsilon = 0.05,
  gene_effects_sd = 1, gene_effect_prob = 0.3, evf_type, randseed = 0,
  plotname = "temp.pdf")
}
\arguments{
\item{ncells}{number of cells}

\item{ngenes}{number of genes}

\item{nevf}{number of evfs}

\item{beta}{a single float value between 0 and 1 determining the bimodality of entire dataset}

\item{epsilon}{a single float value the magnitude of independent noise for each gene in each cell}

\item{nbatches}{number of batches}

\item{evfs}{matrix with ncells rows and nevf columns, can be generated by ContinuousEVF and DiscreteEVF}

\item{gene_effects}{a list of length 3, each element is a sparse matrix determining how each gene is affected by the EVF of that cell}

\item{alphas}{a vector of length ncells}

\item{gcbias}{a vector of length ngenes}

\item{lenbias}{a vector of length ngenes}

\item{batch}{a matrix of ngenes rows and nbatch columns}

\item{batch_id}{a vector of length ncells indicating which batch each cell comes from (randomly generated)}

\item{meta1}{meta data generated from EVF function (population, depth in the tree, etc.)}
}
\description{
Simulate count matrix from evf, gene effects and technical biases
}
