% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SimulateTrueCounts}
\alias{SimulateTrueCounts}
\title{Simulate count matrix from evf, gene effects and technical biases}
\usage{
SimulateTrueCounts(ncells_total, min_popsize = ncells_total, ngenes,
  nevf = 10, evf_type = "one.population", Sigma = 0.3, phyla = NULL,
  gene_effects_sd = 1, gene_effect_prob = 0.3, bimod = 0.3,
  randseed = 0)
}
\arguments{
\item{ngenes}{number of genes}

\item{nevf}{number of evfs}

\item{ncells}{number of cells}

\item{nbatch}{number of batches}

\item{evfs}{matrix with ncells rows and nevf columns, can be generated by ContinuousEVF and DiscreteEVF}

\item{gene_effects}{a list of length 3, each element is a sparse matrix determining how each gene is affected by the EVF of that cell}

\item{alphas}{a vector of length ncells}

\item{beta}{a single float value between 0 and 1 determining the bimodality of entire dataset}

\item{gcbias}{a vector of length ngenes}

\item{lenbias}{a vector of length ngenes}

\item{batch}{a matrix of ngenes rows and nbatch columns}

\item{batch_id}{a vector of length ncells indicating which batch each cell comes from (randomly generated)}

\item{epsilon}{a single float value the magnitude of independent noise for each gene in each cell}
}
\description{
Simulate count matrix from evf, gene effects and technical biases
}
