% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_functions.R
\name{tmm}
\alias{tmm}
\title{TMM normalization}
\usage{
tmm(x)
}
\arguments{
\item{x}{matrix of transcript counts where rows are genes and columns are cells}
}
\value{
list(s,counts) s is a ncells long vector that stores the scaling factors for each cell, counts is the gene-cell matrix after normalization
}
\description{
This function runs the Trimmed Mean of M values (TMM) normalization (from Vallejos2017 supplementary material), TMM trims away extreme log fold changes to normalize the counts based on the remaining set of nondifferentially expressed genes.
}
