% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TrueCounts2Dropped}
\alias{TrueCounts2Dropped}
\title{Sample from true read counts to mimic capture/RT efficiency (only part of the transcripts are represented in the library)}
\usage{
TrueCounts2Dropped(sim_exprs, alpha, alpha_sd)
}
\arguments{
\item{sim_exprs}{the true matrix of transcript counts with ngenes rows and ncells columns}

\item{alpha}{the mean efficiency of capture}

\item{alpha_sd}{the standard deviation of the efficieny of capture for different cells}
}
\value{
dropped_exprs The expression after adjusting for library size (alpha)
}
\description{
This function takes the true transcript counts and samples from it using a binomial distribution
}
\examples{
TrueCounts2Dropped()
}
