% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_functions.R
\name{uq}
\alias{uq}
\title{UQ normalization}
\usage{
uq(x)
}
\arguments{
\item{x}{matrix of transcript counts where rows are genes and columns are cells}
}
\value{
list(s,counts) s is a ncells long vector that stores the scaling factors for each cell, counts is the gene-cell matrix after normalization
}
\description{
This function runs the Upper Quantile (UQ) normalization (from Vallejos2017 supplementary material), scaling factor estimates is proportional to the 75th percentile of the distribution of counts within each cell
}
