% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{BestMatchParams}
\alias{BestMatchParams}
\title{Get the best matched parameter}
\usage{
BestMatchParams(tech, counts, plotfilename, n_optimal = 3,
  depth_range = c(-Inf, Inf), alpha_range = c(-Inf, Inf),
  idx_set = NULL)
}
\arguments{
\item{tech}{'nonUMI','UMI' the match database are constructed based on reasonable values for each technology}

\item{counts}{expression matrix}

\item{plotfilename}{output name for qqplot}

\item{n_optimal}{number of top parameter configurations to return}
}
\value{
three set of best matching parameters that was used to simulate the best matching dataset to the experimental dataset
}
\description{
This function matches a real dataset to a database of summary information of simulated datasets, plots a qqplot for user to visualize similarity between their dataset and the simulated dataset, and suggests parameters to use in the simulation
}
