% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{ContinuousEVF}
\alias{ContinuousEVF}
\title{Generating EVFs for cells sampled along the trajectory of cell development}
\usage{
ContinuousEVF(phyla, ncells, nevf1, nevf2, tip, Sigma, plotting = T, plotname)
}
\arguments{
\item{phyla}{tree for cell developement}

\item{ncells}{number of cells}

\item{nevf1}{Number of EVFs that do not have an impulse signal}

\item{nevf2}{Number of EVFs with an impulse signal}

\item{tip}{The leaf that the path with impulse lead to}

\item{Sigma}{The standard deviation of the brownian motion of EVFs changing along the tree}

\item{plotting}{Whether to plot the trajectory or not}

\item{plotname}{The}
}
\value{
a list of two object, one is the evf, and the other is a dataframe indicating the branch each cell comes from (pop) and its depth in the tree (depth)
}
\description{
Generating EVFs for cells sampled along the trajectory of cell development
}
