% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{ContinuousEVF}
\alias{ContinuousEVF}
\title{Generating EVFs for cells sampled along the trajectory of cell development}
\usage{
ContinuousEVF(phyla, ncells, n_nd_evf, n_de_evf, impulse = F,
  evf_center = 1, vary = "s", Sigma, plotting = T,
  plotname = "cont_evf.pdf", seed)
}
\arguments{
\item{phyla}{tree for cell developement}

\item{ncells}{number of cells}

\item{n_nd_evf}{Number of EVFs that do not have an impulse signal}

\item{n_de_evf}{Number of EVFs with an impulse signal}

\item{impulse}{if the impluse model should be used instead of Brownian motion}

\item{evf_center}{the mean of Gaussain function where the non-Diff EVFs are sampled from}

\item{vary}{which parameters are affected by Diff-EVFs. Can be "kon", "koff", "s", "all", "except_kon", "except_koff", "except_s". Suggestions are "all" or "s"}

\item{Sigma}{The standard deviation of the brownian motion of EVFs changing along the tree}

\item{plotting}{Whether to plot the trajectory or not}

\item{plotname}{The string to be used in the output file name}

\item{seed}{the random seed}

\item{tip}{The leaf that the path with impulse lead to}
}
\value{
a list of two object, one is the evf, and the other is a dataframe indicating the branch each cell comes from (pop) and its depth in the tree (depth)
}
\description{
Generating EVFs for cells sampled along the trajectory of cell development
}
