% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{DE_EVF}
\alias{DE_EVF}
\title{Generating EVFs that is differential in one population and not in all other populations}
\usage{
DE_EVF(de_pop, de_evf_mean, cell_pop, Sigma, nevf, seed)
}
\arguments{
\item{de_pop}{the population that has a differential DE}

\item{de_evf_mean}{the mean of evf of the DE population (mean for all other populations are zero)}

\item{cell_pop}{the population id for each cell from the discrete evf simulation}

\item{Sigma}{The standard deviation of the brownian motion of EVFs changing along the tree}

\item{nevf}{Number of EVFs per cell (taht are DE)}
}
\value{
a list of two object, one is the evf, and the other is a dataframe indicating the population each cell comes from (pop)
}
\description{
Generating EVFs that is differential in one population and not in all other populations
}
