% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{DiscreteEVF}
\alias{DiscreteEVF}
\title{Generating EVFs for cells sampled from tip populations from a tree}
\usage{
DiscreteEVF(phyla, ncells_total, min_popsize, Sigma, nevf)
}
\arguments{
\item{phyla}{tree for cell developement}

\item{ncells_total}{number of cells from all populations}

\item{min_popsize}{size of the rarest population}

\item{Sigma}{The standard deviation of the brownian motion of EVFs changing along the tree}

\item{nevf}{Number of EVFs per cell}
}
\value{
a list of two object, one is the evf, and the other is a dataframe indicating the population each cell comes from (pop)
}
\description{
Generating EVFs for cells sampled from tip populations from a tree
}
