% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{DiscreteEVF}
\alias{DiscreteEVF}
\title{Generating EVFs for cells sampled from tip populations from a tree}
\usage{
DiscreteEVF(phyla, ncells_total, min_popsize, i_minpop, Sigma, n_nd_evf,
  n_de_evf, vary, evf_center, seed)
}
\arguments{
\item{phyla}{tree for cell developement}

\item{ncells_total}{number of cells from all populations}

\item{min_popsize}{size of the rarest population}

\item{i_minpop}{to specify which population has the smallest size}

\item{Sigma}{The standard deviation of the brownian motion of EVFs changing along the tree}

\item{n_nd_evf}{number of non-Diff EVFs}

\item{n_de_evf}{number of Diff EVFs}

\item{vary}{which parameters are affected by Diff-EVFs. Can be "kon", "koff", "s", "all", "except_kon", "except_koff", "except_s". Suggestions are "all" or "s"}

\item{evf_center}{the value used to generated evf means. Suggested value is 1}

\item{seed}{the random seed}
}
\value{
a list of two object, one is the evf, and the other is a dataframe indicating the population each cell comes from (pop)
}
\description{
Generating EVFs for cells sampled from tip populations from a tree
}
