% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{GeneEffects}
\alias{GeneEffects}
\title{Getting GeneEffects matrices}
\usage{
GeneEffects(
  ngenes,
  nevf,
  randseed,
  prob,
  geffect_mean,
  geffect_sd,
  evf_res,
  is_in_module
)
}
\arguments{
\item{ngenes}{number of genes}

\item{nevf}{number of evfs}

\item{prob}{the probability that the effect size is not 0}

\item{geffect_mean}{the mean of the normal distribution where the non-zero effect sizes are dropped from}

\item{geffect_sd}{the standard deviation of the normal distribution where the non-zero effect sizes are dropped from}

\item{evf_res}{the EVFs generated for cells}

\item{is_in_module}{a vector of length ngenes. 0 means the gene is not in any gene module. In case of non-zero values, genes with the same value in this vector are in the same module.}

\item{randomseed}{(should produce same result if ngenes, nevf and randseed are all the same)}
}
\value{
a list of 3 matrices, each of dimension ngenes * nevf
}
\description{
This function randomly generates the effect size of each evf on the dynamic expression parameters
}
