% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{Get_params}
\alias{Get_params}
\title{Getting the parameters for simulating gene expression from EVf and gene effects}
\usage{
Get_params(gene_effects, evf, match_param_den, bimod, scale_s)
}
\arguments{
\item{gene_effects}{a list of three matrices (generated using the GeneEffects function), 
each corresponding to one kinetic parameter. Each matrix has nevf columns, and ngenes rows.}

\item{evf}{a vector of length nevf, the cell specific extrinsic variation factor}

\item{match_param_den}{the fitted parameter distribution density to sample from}

\item{bimod}{the bimodality constant}

\item{scale_s}{a factor to scale the s parameter, which is used to tune the size of the actual cell (small cells have less number of transcripts in total)}
}
\value{
params a matrix of ngenes * 3
}
\description{
This function takes gene_effect and EVF, take their dot product and scale the product to the correct range 
by using first a logistic function and then adding/dividing by constant to their correct range
}
\examples{
Get_params()
}
