% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{multiplot}
\alias{multiplot}
\title{arrange multiple plots from ggplot2 to one figure. 
From http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
arrange multiple plots from ggplot2 to one figure. 
From http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
}
