% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{PlotCountHeatmap}
\alias{PlotCountHeatmap}
\title{Plotting logged expression distribution}
\usage{
PlotCountHeatmap(log_real, mean_counts, given_ord = NA,
  zeropropthres = 0.8, filename, saving = F, data_name)
}
\arguments{
\item{log_real}{the logged distribution of count distribution}

\item{mean_counts}{the average expression for each gene, used for sorting purpose}

\item{given_ord}{the given order of genes}

\item{zeropropthres}{the genes with zeroproportion greater than this number is not plotted (default to 0.8)}

\item{filename}{the name of the output plot}

\item{saving}{if the plot should be saved into a file}

\item{data_name}{a string which is included in the title of the plot to describe the data used}
}
\description{
takes an expression matrix and makes a 2D histogram on the log scale where each row is a gene and the number of samples in a bin is shown by the intensity of the color
}
