% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_functions.R
\name{PlotTsne}
\alias{PlotTsne}
\title{Perform tSNE}
\usage{
PlotTsne(meta, data, evf_type, pca = T, n_pc, perplexity = 30, label,
  saving = F, plotname, system.color = T)
}
\arguments{
\item{meta:}{simulation parameters}

\item{data:}{expression matrix}

\item{plotname:}{the name of the jpeg file}

\item{label:}{the column name of the meta data that the points needs to be colored by}
}
\description{
This function takes the ground truth (meta), the expresssion matrix (data), and performs tSNE, saves the result to a jpeg file with user specified name
}
