% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SimulateTrueCounts}
\alias{SimulateTrueCounts}
\title{Simulate true counts by sampling from a beta-poisson distribution}
\usage{
SimulateTrueCounts(ncells, ngenes, nevf, Sigma = 0.1, phyla, bimod = 0.5,
  gene_effects_mean = 0, gene_effects_sd = 1, gene_effect_prob = 0.3,
  evf_type, randseed = 0)
}
\arguments{
\item{ncells}{number of cells}

\item{ngenes}{number of genes}

\item{nevf}{number of evfs}

\item{nbatch}{number of batches}

\item{evfs}{matrix with ncells rows and nevf columns, can be generated by ContinuousEVF and DiscreteEVF}

\item{gene_effects}{a list of length 3, each element is a sparse matrix determining how each gene is affected by the EVF of that cell}

\item{alphas}{a vector of length ncells}

\item{beta}{a single float value between 0 and 1 determining the bimodality of entire dataset}
}
\description{
This function generates evf, gene effect and true counts
}
