% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SimulateTrueCounts}
\alias{SimulateTrueCounts}
\title{Generate both evf and gene effect and simulate true transcript counts}
\usage{
<<<<<<< HEAD
SimulateTrueCounts(ncells, ngenes, nevf, Sigma = 0.1, phyla, bimod = 0.5,
  gene_effects_mean = 0, gene_effects_sd = 1, gene_effect_prob = 0.3,
  evf_type, randseed = 0)
=======
SimulateTrueCounts(ncells_total, min_popsize, ngenes, nevf = 10,
  evf_type = "one.population", Sigma = 0.3, phyla = NULL,
  gene_effects_sd = 1, gene_effect_prob = 0.3, bimod = 0.3,
  randseed = 0)
>>>>>>> c198b90973e3718e2cbd97d31bffee9e29f0c31c
}
\arguments{
\item{ncells_total}{number of cells}

\item{min_popsize}{the number of cells}

\item{ngenes}{number of genes}

\item{nevf}{number of evfs}

\item{evf_type}{string that is one of the following: 'one.population','discrete','continuous'}

\item{Sigma}{parameter of the std of evf values within the same population}

\item{phyla}{the cell developmental tree if chosing 'discrete' or 'continuous' evf type. Can either be generated randomly or read from newick format file using the ape package}

\item{gene_effect_prob}{the probability that the effect size is not 0}

\item{bimod}{the amount of increased bimodality in the transcript distribution, 0 being not changed from the results calculated using evf and gene effects, and 1 being all genes are bimodal}

\item{randseed}{random seed}

\item{randomseed}{(should produce same result if ngenes, nevf and randseed are all the same)}

\item{gene_effect_sd}{the standard deviation of the normal distribution where the non-zero effect sizes are dropped from}
}
\value{
a list of 4 elements, the first element is true counts, second is cell level meta information, including a matrix of evf and a vector of cell identity, the third is the gene level meta information, and the fourth is the parameters kon, koff and s used to simulation the true counts
}
\description{
Generate both evf and gene effect and simulate true transcript counts
}
