% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{True2ObservedCounts}
\alias{True2ObservedCounts}
\title{Simulate observed count matrix given technical biases and the true counts}
\usage{
True2ObservedCounts(SE = NULL, true_counts, meta_cell, protocol,
  alpha_mean = 0.1, alpha_sd = 0.002, lenslope = 0.02, nbins = 20,
  gene_len, amp_bias_limit = c(-0.2, 0.2), rate_2PCR = 0.8,
  nPCR1 = 16, nPCR2 = 10, LinearAmp = F, LinearAmp_coef = 2000,
  depth_mean, depth_sd, nbatch = 1)
}
\arguments{
\item{SE}{input, should be a summerized experiment rather than a list of elements, default is False}

\item{true_counts}{gene cell matrix}

\item{meta_cell}{the meta information related to cells, will be combined with technical cell level information and returned}

\item{protocol}{a string, can be "nonUMI" or "UMI"}

\item{alpha_mean}{the mean of rate of subsampling of transcripts during capture step, default at 10 percent efficiency}

\item{alpha_sd}{the std of rate of subsampling of transcripts}

\item{lenslope}{amount of length bias}

\item{nbins}{number of bins for gene length}

\item{amp_bias_limit}{range of amplification bias for each gene, a vector of length ngenes}

\item{rate_2PCR}{PCR efficiency, usually very high, default is 0.8}

\item{nPCR1}{the number of PCR cycles, default is 16}

\item{LinearAmp}{if linear amplification is used for pre-amplification step, default is FALSE}

\item{LinearAmp_coef}{the coeficient of linear amplification, that is, how many times each molecule is amplified by}

\item{depth_mean}{mean of sequencing depth}

\item{depth_sd}{std of sequencing depth}

\item{nbatch}{number of batches}

\item{hge2true}{if we add high gene expression to true counts}
}
\description{
Simulate observed count matrix given technical biases and the true counts
}
